refactor(extensions): move color scheme API under src/theme
Per review feedback, color scheme types and declare-function bridge
belong under @apache-superset/core/theme rather than a standalone
colors module. src/colors/index.ts is kept as a re-export shim so
existing @apache-superset/core/colors imports continue to work.
